+++ app/_styles/editorPageStyles.ts
@@ -1,14 +1,3 @@
-export const HowToSaveBtn = {
-  background: "#FDD037",
-  border: "none",
-  borderRadius: "10px",
-  padding: "2px",
-  color: "#000",
-  width: "max-content",
-  fontWeight: "bold",
-  fontSize: "9pt",
-};
-
 export const InputStyle = {
   width: "100%",
   height: "35px",
