+++ app/paint-editor/page.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import PaintEditor from "../_components/paintEditor/PaintEditor";
+import Editor from "../_components/paintEditor/Editor";
 
 const PaintEditorPage = () => {
   return (
     <section className="min-h-full bg-gray-100 flex p-10 justify-center items-start">
-      <PaintEditor />
+      <Editor />
     </section>
   );
 };
